Fixed bug that Mom found
When undoing a move that killed a piece, I didn't actually bring the killed piece back to life because I set the landed on space to NULL no matter what. I fixed my [Board] class's place() function so that it does not place NULL in the space that actually had a piece in it.